refactor: package 구조 domain model별로 구분하여 변경
