+++ src/main/java/subway/domain/line/model/Line.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.line.model;
 
 public class Line {
     private String name;
