+++ src/main/java/subway/station/model/StationRepository.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.station.model;
 
 import java.util.ArrayList;
 import java.util.Collections;
