+++ src/test/java/subway/domain/section/model/SectionTest.java
@@ -1,8 +1,8 @@
-package subway.domain.section;
+package subway.domain.section.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import subway.domain.Station;
+import subway.station.model.Station;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
